backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
PyJWT
passlib[bcrypt]
